pyproject.toml
[tool.poetry]
name = "pushpull"
version = "0.1.0"
description = ""
authors = ["Matthew McAllister <matthew.mcallister.0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.0.1"
requests = "^2.26.0"
pytz = "^2021.1"
psycopg2 = "^2.9.1"
python-dotenv = "^0.20.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.4"}

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"
mypy = "^0.971"
types-requests = "^2.28.9"
types-pytz = "^2022.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
